+++ src/main/java/com/nastya/images/dao/TitleImageDao.java
@@ -3,6 +3,6 @@ package com.nastya.images.dao;
 import com.nastya.images.entity.TitleImage;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TitleImageDao extends JpaRepository<TitleImage, Short> {
+public interface TitleImageDao extends JpaRepository<TitleImage, Integer> {
 
 }
